Consolidate component files (stories, tests, components) into subfolders inside . This will be the folder structure moving forward
