+++ components/button/Button.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { fn } from "@storybook/test";
-import { Button } from '../../components/ui/Button';
+import { Button } from './Button';
 
 //👇 This default export determines where your story goes in the story list
 const meta: Meta<typeof Button> = {
