+++ components/button/Button.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
-import { Button } from "@/components/ui/Button";
+import { Button } from "./Button";
 
 it('should render a button with no variant defined', () => {
   render(<Button />)
